+++ CHANGELOG.md
@@ -23,6 +23,9 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/virtual-users/branches/compare/master%0Drelease-3.3.0
 
+### Fixed
+- Add missing browser parameter when serialising VirtualUserOptions to command line parameters.
+
 ## [3.3.0] - 2018-11-27
 [3.3.0]: https://bitbucket.org/atlassian/virtual-users/branches/compare/release-3.3.0%0Drelease-3.2.0